backend/requirements.txt
# Backend requirements for multi-agent system

langchain-groq
langchain-tavily
langchain-core
fastapi
uvicorn[standard]
langchain
langgraph
python-dotenv
pydantic
pydantic-settings
PyYAML
websockets
httpx